backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.50
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.2
anthropic==0.39.0
tenacity==8.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
aiofiles==23.2.1
